requirements.txt
streamlit
google-generativeai
fpdf
python-docx
pandas
openpyxl
